Initial work on handling text formatting transparently...
This took a lot of fiddling to meet all my criteria! However, a user should
be able to use blessings.Terminal like syntax for all the content in their
UI elements, and we should just be able to work out a) the size of the
content appropriately (i.e. not include all the escape sequences in length
calculations) and b) make sure that 'normal' is relative to what the parent
element defines as 'normal', rather than always what the terminal defines.

Note, there are some bits here that seem quite inefficient and will
probably need to be refined. I'm going to do a bunch more unit testing and
some examples first though, so that I can check all is hunky-dory!