Added in Input display element which can take and display text input...
it calls back to its associated root element when it's updated so that the UI
is redrawn. This is a *work in progress*! A number of things I want to change
include marking the element as needing update, and then calling update on the
root element, perhaps automatically. (I do want to keep everything event-driven
as much as possible.) I also want to make the Input element have the kind of
placeholder hint text that you get on web forms as an option .