Made element drawing operations return list of new Block objects...
I'd ummed and ahhed about this a while ago, but implemented it now because it
was seeming silly to pass a Terminal object all the way down from the root
element just to get the children to call it back to draw themselves. Now, the
children provide all the information required up to root, which takes care of
the drawing.

This also means that it's much nicer to unit test, becuause we can intercept
the draw information, rather than having to mock out stuff to catch
side-effects.